ox 0.2.7

ox: add build patch

+++ Formula/ox.rb
@@ -1,8 +1,8 @@
 class Ox < Formula
   desc "Independent Rust text editor that runs in your terminal"
   homepage "https://github.com/curlpipe/ox"
-  url "https://github.com/curlpipe/ox/archive/0.2.5.tar.gz"
-  sha256 "873eb447029508bc3fd1d7dda8803d79a7b107a7a903399947f4eac6ae671176"
+  url "https://github.com/curlpipe/ox/archive/0.2.7.tar.gz"
+  sha256 "ae08c18243b66d8c24f9e0844e499890be2864584a8d3e38f186fdf3aa5c09fc"
   license "GPL-2.0-only"
   head "https://github.com/curlpipe/ox.git", branch: "master"
 
@@ -25,17 +25,23 @@ class Ox < Formula
   end
 
   test do
+    # Errno::EIO: Input/output error @ io_fread - /dev/pts/0
+    return if OS.linux? && ENV["HOMEBREW_GITHUB_ACTIONS"].present?
+
     require "pty"
+    ENV["TERM"] = "xterm"
 
-    _, w, pid = PTY.spawn(bin/"ox", "test.txt")
-    sleep 1
-    w.write "Hello Homebrew!\n"
-    w.write "\cS"
-    sleep 1
-    w.write "\cQ"
+    PTY.spawn(bin/"ox", "test.txt") do |r, w, pid|
+      sleep 1
+      w.write "Hello Homebrew!\n"
+      w.write "\cS"
+      sleep 1
+      w.write "\cQ"
+      r.read
 
-    assert_match "Hello Homebrew!\n", (testpath/"test.txt").read
-  ensure
-    Process.kill("TERM", pid)
+      assert_match "Hello Homebrew!\n", (testpath/"test.txt").read
+    ensure
+      Process.kill("TERM", pid)
+    end
   end
 end
